refactor(cli): extract maintenance actions
